+++ mkdocs/commands/new.py
@@ -1,7 +1,3 @@
-# coding: utf-8
-from __future__ import unicode_literals
-
-import io
 import logging
 import os
 
@@ -43,7 +39,7 @@ def new(output_dir):
         os.mkdir(output_dir)
 
     log.info('Writing config file: %s', config_path)
-    io.open(config_path, 'w', encoding='utf-8').write(config_text)
+    open(config_path, 'w', encoding='utf-8').write(config_text)
 
     if os.path.exists(index_path):
         return
@@ -51,4 +47,4 @@ def new(output_dir):
     log.info('Writing initial docs: %s', index_path)
     if not os.path.exists(docs_dir):
         os.mkdir(docs_dir)
-    io.open(index_path, 'w', encoding='utf-8').write(index_text)
+    open(index_path, 'w', encoding='utf-8').write(index_text)