+++ mkdocs/contrib/search/search_index.py
@@ -1,7 +1,3 @@
-# coding: utf-8
-
-from __future__ import unicode_literals
-
 import os
 import re
 import json
@@ -10,17 +6,12 @@ import subprocess
 
 from lunr import lunr
 
-from mkdocs import utils
-
-try:                                    # pragma: no cover
-    from html.parser import HTMLParser  # noqa
-except ImportError:                     # pragma: no cover
-    from HTMLParser import HTMLParser   # noqa
+from html.parser import HTMLParser
 
 log = logging.getLogger(__name__)
 
 
-class SearchIndex(object):
+class SearchIndex:
     """
     Search index is a collection of pages and sections (heading
     tags and their following content are sections).
@@ -52,7 +43,7 @@ class SearchIndex(object):
 
         self._entries.append({
             'title': title,
-            'text': utils.text_type(text.encode('utf-8'), encoding='utf-8'),
+            'text': str(text.encode('utf-8'), encoding='utf-8'),
             'location': loc
         })
 
@@ -96,7 +87,7 @@ class SearchIndex(object):
         if toc_item is not None:
             self._add_entry(
                 title=toc_item.title,
-                text=u" ".join(section.text),
+                text=" ".join(section.text),
                 loc=abs_url + toc_item.url
             )
 
@@ -125,7 +116,7 @@ class SearchIndex(object):
                     log.debug('Pre-built search index created successfully.')
                 else:
                     log.warning('Failed to pre-build search index. Error: {}'.format(err))
-            except (OSError, IOError, ValueError) as e:
+            except (OSError, ValueError) as e:
                 log.warning('Failed to pre-build search index. Error: {}'.format(e))
         elif self.config['prebuild_index'] == 'python':
             idx = lunr(
@@ -151,9 +142,7 @@ class HTMLStripper(HTMLParser):
     """
 
     def __init__(self, *args, **kwargs):
-        # HTMLParser is a old-style class in Python 2, so
-        # super() wont work here.
-        HTMLParser.__init__(self, *args, **kwargs)
+        super().__init__(*args, **kwargs)
 
         self.data = []
 
@@ -167,7 +156,7 @@ class HTMLStripper(HTMLParser):
         return '\n'.join(self.data)
 
 
-class ContentSection(object):
+class ContentSection:
     """
     Used by the ContentParser class to capture the information we
     need when it is parsing the HMTL.
@@ -195,9 +184,7 @@ class ContentParser(HTMLParser):
 
     def __init__(self, *args, **kwargs):
 
-        # HTMLParser is a old-style class in Python 2, so
-        # super() wont work here.
-        HTMLParser.__init__(self, *args, **kwargs)
+        super().__init__(*args, **kwargs)
 
         self.data = []
         self.section = None