+++ mkdocs/tests/new_tests.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python
-# coding: utf-8
 
-from __future__ import unicode_literals
 import tempfile
 import unittest
 import os